src/main/java/me/savchik/statistics/controller/StatisticsController.java
package me.savchik.statistics.controller;

import me.savchik.statistics.entity.Statistics;
import me.savchik.statistics.service.StatisticsService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/statistics")
public class StatisticsController {

    private StatisticsService service;

    @Autowired
    public StatisticsController(StatisticsService service) {
        this.service = service;
    }

    @GetMapping
    public Statistics getStatistics() {
        return service.getStatistics();
    }

}
